sys/ximage/gstximagesrc.c: Fix presumably copy'n'pasto for 16bpp depth.
Original commit message from CVS:
* sys/ximage/gstximagesrc.c: (composite_pixel):
Fix presumably copy'n'pasto for 16bpp depth.

+++ sys/ximage/gstximagesrc.c
@@ -265,7 +265,7 @@ composite_pixel (GstXContext * xcontext, guchar * dest, guchar * src)
       color = *dest;
       break;
     case 16:
-      color = GUINT16_FROM_LE (*(guint32 *) (dest));
+      color = GUINT16_FROM_LE (*(guint16 *) (dest));
       break;
     case 32:
       color = GUINT32_FROM_LE (*(guint32 *) (dest));
